pyproject.toml
[tool.poetry]
name = "sosec-digital-twins"
version = "0.1.0"
description = ""
authors = ["Miriam <91884702+miriamcoccia@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.36.0"
pyyaml = "^6.0.1"
pydantic = "^2.8.2"
pandas = "^2.2.2"
tqdm = "^4.66.4"
eventregistry = "^9.1"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
